requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.3.0
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-debug-toolbar==4.3.0
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.0
drf-spectacular==0.27.2
future==1.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
serializers==0.2.4
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
